+++ examples/all_attributes.rs
@@ -1,5 +1,6 @@
 use bevy::prelude::*;
-use bevy_mod_sysfail::*;
+use bevy_mod_sysfail::prelude::*;
+use bevy_mod_sysfail::Dedup;
 
 use thiserror::Error;
 
@@ -10,6 +11,12 @@ struct Foo;
 enum GizmoError {
     #[error("A Gizmo error")]
     Error,
+}
+
+impl Dedup for GizmoError {
+    type ID = ();
+
+    fn identify(&self) {}
 }
 
 fn main() {
@@ -24,25 +31,23 @@ fn drag_gizmo(time: Res<Time>) {
     println!("drag time is: {}", time.elapsed_seconds());
     let _ = Err(GizmoError::Error)?;
     println!("This will never print");
-    Ok(())
 }
 
-#[sysfail]
-fn place_gizmo() -> Result<(), Log<&'static str, Info>> {
+#[sysfail(Log<&'static str, Info>)]
+fn place_gizmo() {
     let () = Result::<(), &'static str>::Ok(())?;
     println!("this line should actually show up");
     let _ = Err("Ah, some creative use of info logging I see")?;
-    Ok(())
 }
 
 /// This also has some doc
-#[quick_sysfail]
+#[sysfail(Ignore)]
 fn delete_gizmo(time: Res<Time>, mut query: Query<&mut Transform>, foos: Query<Entity, With<Foo>>) {
     println!("delete time is: {}", time.elapsed_seconds());
     for foo in &foos {
-        let mut trans = query.get_mut(foo).ok()?;
+        let mut trans = query.get_mut(foo)?;
         trans.translation += Vec3::Y;
     }
-    let _ = None?;
+    let _ = Err(())?;
     println!("This will never print");
 }